fix error at ex012 (change int to float)

+++ workouts/ex012.py
@@ -1,6 +1,6 @@
 # Faça um algorítmo que receba um preço de um produto e calcule seu preço com desconto de 5%  # noqa: E501
 
-value = int(input('Digite o valor do produto: '))
+value = float(input('Digite o valor do produto: '))
 discount = ((5 / 100) * value)
 newValue = value - discount
